pyproject.toml
[project]
name = "chtc-memory-analyzer"
version = "0.2.0"
description = "Analyze HTCondor job history for memory usage patterns and identify resource over-allocation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ian Ross", email = "iaross@wisc.edu"}
]
keywords = ["htcondor", "htc", "memory", "analysis", "resource-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    "htcondor>=24.0.0",
    "click>=8.0.0",
    "pandas>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "ruff>=0.7.0",
    "ty",
    "pre-commit>=3.0.0",
]
elasticsearch = [
    "elasticsearch>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/iross/chtc-memory-analyzer"
Repository = "https://github.com/iross/chtc-memory-analyzer"
Issues = "https://github.com/iross/chtc-memory-analyzer/issues"

[project.scripts]
chtc-analyze = "chtc_memory_analyzer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chtc_memory_analyzer"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []
